Restructure source module
Implements #52 and flattens the structure a bit like suggested in #49
